# openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: K8sPenTool-ng API
  description: |
    K8sPenTool-ng - Kubernetes Container Penetration Testing Platform

    Integrates CDK (Container Defense Kit) exploits with AI-driven attack chain automation.
    All endpoints use JSON request/response. Real-time events via WebSocket at `/api/v1/ws`.
  version: "2.0.0"
  contact:
    name: K8sPenTool Team
    url: https://github.com/trymonoly/K8sPenTool-ng
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /dashboard/probe:
    post:
      tags: [Access]
      summary: Probe Dashboard accessibility and auth bypass indicators
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardProbeRequest'
      responses:
        '200':
          description: Dashboard probe results
...
components:
  schemas:
...
    DashboardProbeRequest:
      type: object
      required: [target_host]
      properties:
        target_host:
          type: string
        dashboard_port:
          type: integer
        dashboard_path:
          type: string
        use_https:
          type: boolean
        timeout_sec:
          type: integer
        skip_tls:
          type: boolean
...